plagiarism_turnitin: Cast maxsubmissionsizebytes value to bigint
Can be too large for an int, which causes cron to error out.

+++ lib.php
@@ -2392,7 +2392,7 @@ class plagiarism_plugin_turnitin extends plagiarism_plugin {
                        WHERE apc.plugin = ?
                          AND apc.subtype = ?
                          AND apc.name = ?
-                         AND (CAST(apc.value AS int) > ?
+                         AND (CAST(apc.value AS bigint) > ?
                           OR apc.value = ?)";
 
         $selectparams = array(
